+++ src/main/java/com/jasonvillar/works/register/user/UserController.java
@@ -1,16 +1,21 @@
 package com.jasonvillar.works.register.user;
 
+import com.jasonvillar.works.register.authentication.SecurityUserDetailsService;
 import com.jasonvillar.works.register.user.port.in.AddAdminRoleToUserRequest;
+import com.jasonvillar.works.register.user.port.in.ChangePasswordRequest;
 import com.jasonvillar.works.register.user.port.in.UserRequestAdapter;
 import com.jasonvillar.works.register.user.port.out.UserDTOAdapter;
 import com.jasonvillar.works.register.user.port.in.UserRequest;
 import com.jasonvillar.works.register.user.port.out.UserDTO;
+import io.swagger.v3.oas.annotations.security.SecurityRequirements;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -29,6 +34,8 @@ public class UserController {
     private final UserRequestAdapter userRequestAdapter;
 
     private final UserDTOAdapter userDTOAdapter;
+
+    private final SecurityUserDetailsService securityUserDetailsService;
 
     @Secured("View users")
     @GetMapping(value = "/user/{id}", produces = "application/json")
@@ -114,6 +121,43 @@ public class UserController {
             return new ResponseEntity<>("Added admin role to user with id " + request.id(), HttpStatus.CREATED);
         } else {
             return new ResponseEntity<>("An error has occurred", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @SecurityRequirements
+    @GetMapping(value = "/user/validate/name/{name}/email/{email}/code/{code}")
+    public ResponseEntity<Object> validateUser(@PathVariable String name, @PathVariable String email, @PathVariable String code) {
+        Optional<User> userOptional = this.service.getOptionalByNameAndEmailAndValidatedAndCode(name, email, false, code);
+
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
+
+            user.setCode(null);
+            user.setValidated(true);
+
+            user = service.save(user);
+            UserDTO userDTO = userDTOAdapter.apply(user);
+            return ResponseEntity.ok(userDTO);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
+    @PostMapping(value = "/user/change-password")
+    public ResponseEntity<String> resetPassword(@AuthenticationPrincipal UserDetails userDetails, @RequestBody ChangePasswordRequest changePasswordRequest) {
+        if (!changePasswordRequest.newPassword().equals(changePasswordRequest.repeatNewPassword())) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("The new password has differences");
+        }
+
+        long userId = this.securityUserDetailsService.getAuthenticatedUserId(userDetails);
+        User user = this.service.getById(userId);
+
+        if (this.service.passwordMatchWithActual(changePasswordRequest.oldPassword(), user.getPassword())) {
+            user.setPassword(this.service.plainPasswordToBcrypt(changePasswordRequest.newPassword()));
+            this.service.save(user);
+            return ResponseEntity.ok("The password has been changed");
+        } else {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("The actual password is incorrect");
         }
     }
 }