+++ src/test/java/com/jasonvillar/works/register/unit/user/user_not_validated/UserNotValidatedServiceTest.java
@@ -1,5 +1,6 @@
 package com.jasonvillar.works.register.unit.user.user_not_validated;
 
+import com.jasonvillar.works.register.email.EmailService;
 import com.jasonvillar.works.register.user.user_not_validated.UserNotValidated;
 import com.jasonvillar.works.register.user.user_not_validated.UserNotValidatedId;
 import com.jasonvillar.works.register.user.user_not_validated.UserNotValidatedRepository;
@@ -7,6 +8,8 @@ import com.jasonvillar.works.register.user.user_not_validated.UserNotValidatedSe
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
@@ -14,12 +17,16 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
 
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 
 @ExtendWith(MockitoExtension.class)
 class UserNotValidatedServiceTest {
     @Mock
     private UserNotValidatedRepository userNotValidatedRepository;
+
+    @Mock
+    private EmailService emailService;
 
     @InjectMocks
     private UserNotValidatedService userNotValidatedService;
@@ -29,24 +36,6 @@ class UserNotValidatedServiceTest {
         String code = this.userNotValidatedService.makeRandomValidationCode();
         Assertions.assertThat(code).hasSize(6);
         Assertions.assertThat(Integer.valueOf(code).getClass()).hasSameClassAs(Integer.class);
-    }
-
-    @Test
-    void makeValidationCodeForUserNotValidated() {
-        UserNotValidatedId userNotValidatedId = UserNotValidatedId.builder()
-                .name("Test User Name")
-                .email("test@test.com")
-                .build();
-
-        UserNotValidated userNotValidated = UserNotValidated.builder()
-                .userNotValidatedId(userNotValidatedId)
-                .password("sarasa")
-                .build();
-
-        userNotValidated = this.userNotValidatedService.makeValidationCodeForUserNotValidated(userNotValidated);
-
-        Assertions.assertThat(userNotValidated.getCode()).isNotEmpty();
-        Assertions.assertThat(userNotValidated.getCode()).isNotNull();
     }
 
     @Test
@@ -125,5 +114,30 @@ class UserNotValidatedServiceTest {
         boolean deleted = this.userNotValidatedService.deleteById(userNotValidatedId);
 
         Assertions.assertThat(deleted).isTrue();
+
+        Mockito.when(this.userNotValidatedRepository.deleteByUserNotValidatedId(userNotValidatedId)).thenReturn(0L);
+
+        deleted = this.userNotValidatedService.deleteById(userNotValidatedId);
+
+        Assertions.assertThat(deleted).isFalse();
+    }
+
+    @Test
+    void sendValidationCode() {
+        Mockito.when(this.emailService.sendSimpleMessage(anyString(), anyString(), anyString())).thenReturn(true);
+        boolean send = this.userNotValidatedService.sendValidationCode("Admin", "test@test.com", "123456", null, true);
+        Assertions.assertThat(send).isTrue();
+
+        Mockito.when(this.emailService.sendSimpleMessage(anyString(), anyString(), anyString())).thenReturn(true);
+        send = this.userNotValidatedService.sendValidationCode("Admin", "test@test.com", "123456", null, false);
+        Assertions.assertThat(send).isTrue();
+
+        Mockito.when(this.emailService.sendSimpleMessage(anyString(), anyString(), anyString())).thenReturn(true);
+        send = this.userNotValidatedService.sendValidationCode("Admin", "test@test.com", "123456", "http://frontend/code", true);
+        Assertions.assertThat(send).isTrue();
+
+        Mockito.when(this.emailService.sendSimpleMessage(anyString(), anyString(), anyString())).thenReturn(true);
+        send = this.userNotValidatedService.sendValidationCode("Admin", "test@test.com", "123456", "http://frontend/code", false);
+        Assertions.assertThat(send).isTrue();
     }
 }